pulp-devel/requirements.txt
flake8
pyyaml
